Fixes #32, update to xhr2 with integration tests
