+++ ampersand-sync.js
@@ -110,8 +110,8 @@ module.exports = function (method, model, options) {
     // Make the request. The callback executes functions that are compatible
     // With jQuery.ajax's syntax.
     var request = options.xhr = options.xhrImplementation(ajaxSettings, function (err, resp, body) {
-        if (err) {
-            if (options.error) return options.error(resp, 'error', err.message);
+        if (err || resp.statusCode >= 400) {
+            if (options.error) return options.error(resp, 'error', err.message || body);
         } else {
             // Parse body as JSON if a string.
             if (body && typeof body === 'string') {
